update GUI, final push
