+++ README.md
@@ -1,1 +1,8 @@
 # ist_303_group_project Group 4
+
+
+Software Development Project
+
+To run : python sd_driver_gui.py and the main menu will appear. 
+
+(Dependencies: Python 3, SQLite, PyQt5, Pickle)
