+++ sd_2nd_band_assign.py
@@ -8,15 +8,15 @@ Check if dorm pickle files exists. If so, load the file
 
 """
 
-if (os.path.exists("band_assignments_June.p")):
+if (os.path.exists("sec_Male_band_assignments_June.p")):
 	
-	band_assignments_June = pickle.load(open("band_assignments_June.p", "rb"))
+	sec_Male_band_assignments_June = pickle.load(open("sec_Male_band_assignments_June.p", "rb"))
 
 else:
+
+	sec_Male_band_assignments_June = {
 	
-	band_assignments_June = {
-	
-		'Band1' : {
+		'Boy_Band1' : {
 			'Singer' : [],
 			'Guitarist' : [],
 			'Drummer' : [],
@@ -25,7 +25,7 @@ else:
 			'Instrumentalist' : []
 		},
 
-		'Band2' : {
+		'Boy_Band2' : {
 			'Singer' : [],
 			'Guitarist' : [],
 			'Drummer' : [],
@@ -34,7 +34,7 @@ else:
 			'Instrumentalist' : []
 		},
 
-		'Band3' : {
+		'Boy_Band3' : {
 			'Singer' : [],
 			'Guitarist' : [],
 			'Drummer' : [],
@@ -43,43 +43,7 @@ else:
 			'Instrumentalist' : []
 		},
 
-		'Band4' : {
-			'Singer' : [],
-			'Guitarist' : [],
-			'Drummer' : [],
-			'Bassist' : [],
-			'Keyboardist' : [],
-			'Instrumentalist' : []
-		},
-
-		'Band5' : {
-			'Singer' : [],
-			'Guitarist' : [],
-			'Drummer' : [],
-			'Bassist' : [],
-			'Keyboardist' : [],
-			'Instrumentalist' : []
-		},
-
-		'Band6' : {
-			'Singer' : [],
-			'Guitarist' : [],
-			'Drummer' : [],
-			'Bassist' : [],
-			'Keyboardist' : [],
-			'Instrumentalist' : []
-		},
-
-		'Band7' : {
-			'Singer' : [],
-			'Guitarist' : [],
-			'Drummer' : [],
-			'Bassist' : [],
-			'Keyboardist' : [],
-			'Instrumentalist' : []
-		},
-
-		'Band8' : {
+		'Boy_Band4' : {
 			'Singer' : [],
 			'Guitarist' : [],
 			'Drummer' : [],
@@ -88,17 +52,19 @@ else:
 			'Instrumentalist' : []
 		}
 
-	} 
 
-if (os.path.exists("band_assignments_July.p")):
+	}
+
+
+if (os.path.exists("sec_Female_band_assignments_June.p")):
 	
-	band_assignments_July = pickle.load(open("band_assignments_July.p", "rb"))
+	sec_Female_band_assignments_June = pickle.load(open("sec_Female_band_assignments_June.p", "rb"))
 
 else:
+
+	sec_Female_band_assignments_June = {
 	
-	band_assignments_July = {
-	
-		'Band1' : {
+		'Girl_Band1' : {
 			'Singer' : [],
 			'Guitarist' : [],
 			'Drummer' : [],
@@ -107,7 +73,7 @@ else:
 			'Instrumentalist' : []
 		},
 
-		'Band2' : {
+		'Girl_Band2' : {
 			'Singer' : [],
 			'Guitarist' : [],
 			'Drummer' : [],
@@ -116,7 +82,7 @@ else:
 			'Instrumentalist' : []
 		},
 
-		'Band3' : {
+		'Girl_Band3' : {
 			'Singer' : [],
 			'Guitarist' : [],
 			'Drummer' : [],
@@ -125,43 +91,7 @@ else:
 			'Instrumentalist' : []
 		},
 
-		'Band4' : {
-			'Singer' : [],
-			'Guitarist' : [],
-			'Drummer' : [],
-			'Bassist' : [],
-			'Keyboardist' : [],
-			'Instrumentalist' : []
-		},
-
-		'Band5' : {
-			'Singer' : [],
-			'Guitarist' : [],
-			'Drummer' : [],
-			'Bassist' : [],
-			'Keyboardist' : [],
-			'Instrumentalist' : []
-		},
-
-		'Band6' : {
-			'Singer' : [],
-			'Guitarist' : [],
-			'Drummer' : [],
-			'Bassist' : [],
-			'Keyboardist' : [],
-			'Instrumentalist' : []
-		},
-
-		'Band7' : {
-			'Singer' : [],
-			'Guitarist' : [],
-			'Drummer' : [],
-			'Bassist' : [],
-			'Keyboardist' : [],
-			'Instrumentalist' : []
-		},
-
-		'Band8' : {
+		'Girl_Band4' : {
 			'Singer' : [],
 			'Guitarist' : [],
 			'Drummer' : [],
@@ -170,17 +100,19 @@ else:
 			'Instrumentalist' : []
 		}
 
-	} 
 
-if (os.path.exists("band_assignments_August.p")):
+	}
+
+
+if (os.path.exists("sec_Male_band_assignments_July.p")):
 	
-	band_assignments_August = pickle.load(open("band_assignments_August.p", "rb"))
+	sec_Male_band_assignments_July = pickle.load(open("sec_Male_band_assignments_July.p", "rb"))
 
 else:
+
+	sec_Male_band_assignments_July = {
 	
-	band_assignments_August = {
-	
-		'Band1' : {
+		'Boy_Band1' : {
 			'Singer' : [],
 			'Guitarist' : [],
 			'Drummer' : [],
@@ -189,7 +121,7 @@ else:
 			'Instrumentalist' : []
 		},
 
-		'Band2' : {
+		'Boy_Band2' : {
 			'Singer' : [],
 			'Guitarist' : [],
 			'Drummer' : [],
@@ -198,7 +130,7 @@ else:
 			'Instrumentalist' : []
 		},
 
-		'Band3' : {
+		'Boy_Band3' : {
 			'Singer' : [],
 			'Guitarist' : [],
 			'Drummer' : [],
@@ -207,43 +139,7 @@ else:
 			'Instrumentalist' : []
 		},
 
-		'Band4' : {
-			'Singer' : [],
-			'Guitarist' : [],
-			'Drummer' : [],
-			'Bassist' : [],
-			'Keyboardist' : [],
-			'Instrumentalist' : []
-		},
-
-		'Band5' : {
-			'Singer' : [],
-			'Guitarist' : [],
-			'Drummer' : [],
-			'Bassist' : [],
-			'Keyboardist' : [],
-			'Instrumentalist' : []
-		},
-
-		'Band6' : {
-			'Singer' : [],
-			'Guitarist' : [],
-			'Drummer' : [],
-			'Bassist' : [],
-			'Keyboardist' : [],
-			'Instrumentalist' : []
-		},
-
-		'Band7' : {
-			'Singer' : [],
-			'Guitarist' : [],
-			'Drummer' : [],
-			'Bassist' : [],
-			'Keyboardist' : [],
-			'Instrumentalist' : []
-		},
-
-		'Band8' : {
+		'Boy_Band4' : {
 			'Singer' : [],
 			'Guitarist' : [],
 			'Drummer' : [],
@@ -252,98 +148,220 @@ else:
 			'Instrumentalist' : []
 		}
 
-	} 
-"""
-Calculate the Gender Distribution of a given Band.  
 
-gender : camper gender to get distribution of 
-temp_dict : the dictionary that represents a given Band 
-
-Returns 1 if the gender distribution is responable for a 
-camper of 'gender' to be inserted into the Band
-
-Returns 0 Otherwise 
-
-"""
-def get_gender_distribution(gender, temp_dict):
-
-	male_count = 0 
-	female_count = 0
-
-	for k in temp_dict:
-		if len(temp_dict[k]) != 0:
-			if( temp_dict[k][0][1] == "Male"):
-				male_count += 1
-			else:
-				female_count += 1
-
-	if gender == "Female" and female_count < 3:
-		return 1
-	elif gender == "Male" and male_count < 3:
-		return 1
-	else: 
-		return 0 
+	}
 
 
-"""
-Assigns Band to a Camper
+if (os.path.exists("sec_Female_band_assignments_July.p")):
+	
+	sec_Female_band_assignments_July = pickle.load(open("sec_Female_band_assignments_July.p", "rb"))
 
-id : Camper ID
-band_role : Camper's proposed band role 
-gender : Camper's Gender
+else:
 
-"""
-def assign_band(id, band_role, gender, rank, session, avoid_band=None):
+	sec_Female_band_assignments_July = {
+	
+		'Girl_Band1' : {
+			'Singer' : [],
+			'Guitarist' : [],
+			'Drummer' : [],
+			'Bassist' : [],
+			'Keyboardist' : [],
+			'Instrumentalist' : []
+		},
 
-	band_filename = "band_assignments_"+session+".p"
+		'Girl_Band2' : {
+			'Singer' : [],
+			'Guitarist' : [],
+			'Drummer' : [],
+			'Bassist' : [],
+			'Keyboardist' : [],
+			'Instrumentalist' : []
+		},
 
-	if (session == "June"):
-		band_assignments = band_assignments_June
-	elif (session == "July"):
-		band_assignments = band_assignments_July
+		'Girl_Band3' : {
+			'Singer' : [],
+			'Guitarist' : [],
+			'Drummer' : [],
+			'Bassist' : [],
+			'Keyboardist' : [],
+			'Instrumentalist' : []
+		},
+
+		'Girl_Band4' : {
+			'Singer' : [],
+			'Guitarist' : [],
+			'Drummer' : [],
+			'Bassist' : [],
+			'Keyboardist' : [],
+			'Instrumentalist' : []
+		}
+
+
+	}
+
+if (os.path.exists("sec_Male_band_assignments_August.p")):
+	
+	sec_Male_band_assignments_August = pickle.load(open("sec_Male_band_assignments_August.p", "rb"))
+
+else:
+
+	sec_Male_band_assignments_August = {
+	
+		'Boy_Band1' : {
+			'Singer' : [],
+			'Guitarist' : [],
+			'Drummer' : [],
+			'Bassist' : [],
+			'Keyboardist' : [],
+			'Instrumentalist' : []
+		},
+
+		'Boy_Band2' : {
+			'Singer' : [],
+			'Guitarist' : [],
+			'Drummer' : [],
+			'Bassist' : [],
+			'Keyboardist' : [],
+			'Instrumentalist' : []
+		},
+
+		'Boy_Band3' : {
+			'Singer' : [],
+			'Guitarist' : [],
+			'Drummer' : [],
+			'Bassist' : [],
+			'Keyboardist' : [],
+			'Instrumentalist' : []
+		},
+
+		'Boy_Band4' : {
+			'Singer' : [],
+			'Guitarist' : [],
+			'Drummer' : [],
+			'Bassist' : [],
+			'Keyboardist' : [],
+			'Instrumentalist' : []
+		}
+
+
+	}
+
+
+if (os.path.exists("sec_Female_band_assignments_August.p")):
+	
+	sec_Female_band_assignments_August = pickle.load(open("sec_Female_band_assignments_August.p", "rb"))
+
+else:
+
+	sec_Female_band_assignments_August = {
+	
+		'Girl_Band1' : {
+			'Singer' : [],
+			'Guitarist' : [],
+			'Drummer' : [],
+			'Bassist' : [],
+			'Keyboardist' : [],
+			'Instrumentalist' : []
+		},
+
+		'Girl_Band2' : {
+			'Singer' : [],
+			'Guitarist' : [],
+			'Drummer' : [],
+			'Bassist' : [],
+			'Keyboardist' : [],
+			'Instrumentalist' : []
+		},
+
+		'Girl_Band3' : {
+			'Singer' : [],
+			'Guitarist' : [],
+			'Drummer' : [],
+			'Bassist' : [],
+			'Keyboardist' : [],
+			'Instrumentalist' : []
+		},
+
+		'Girl_Band4' : {
+			'Singer' : [],
+			'Guitarist' : [],
+			'Drummer' : [],
+			'Bassist' : [],
+			'Keyboardist' : [],
+			'Instrumentalist' : []
+		}
+
+
+	}
+
+
+
+
+def assign_sec_band(id, band_role, gender, rank, session, avoid_band=None):
+	
+	band_filename = "sec_"+gender+"_band_assignments_"+session+".p"
+
+	if (gender == "Male"):
+		if (session == "June"):
+			band_assignments = sec_Male_band_assignments_June
+		elif (session == "July"):
+			band_assignments = sec_Male_band_assignments_July
+		else:
+			band_assignments = sec_Male_band_assignments_August
 	else:
-		band_assignments = band_assignments_August
+		if (session == "June"):
+			band_assignments = sec_Female_band_assignments_June
+		elif (session == "July"):
+			band_assignments = sec_Female_band_assignments_July
+		else:
+			band_assignments = sec_Female_band_assignments_August
 
 	for key in band_assignments:
 		if (len(band_assignments[key][band_role]) == 0 and key != avoid_band):
 
 			temp_dict = band_assignments[key]
 			rank_count = 0
+
 			for i in temp_dict:
 				if (len(temp_dict[i]) == 1):
 					curr_rank = temp_dict[i][0][2]
 					if curr_rank != "None": rank_count += 1
 
+			if (rank != "None" and rank_count < 4):
+				band_assignments[key][band_role].append((id, gender, rank))
+				pickle.dump( band_assignments, open(band_filename, "wb") )
+				# returns band name
+				return key
+			elif (rank == "None"):
+				band_assignments[key][band_role].append((id, gender, rank))
+				pickle.dump( band_assignments, open(band_filename, "wb") )
+				# returns band name
+				return key
+			else:
+				continue 
 
-			if(get_gender_distribution(gender, temp_dict)):
-				if (rank != "None" and rank_count < 4):
-					band_assignments[key][band_role].append((id, gender, rank))
-					pickle.dump( band_assignments, open(band_filename, "wb") )
-					# returns band name
-					return key 
-				elif (rank == "None"):
-					band_assignments[key][band_role].append((id, gender, rank))
-					pickle.dump( band_assignments, open(band_filename, "wb") )
-					# returns band name
-					return key
-				else:
-					continue
-
-	return "Error"  
-
-				
+	return "Error"
+		
 
 """
 Obtain the band of a Camper of given 'id' and 'band_role'
 """
-def get_band_name(id, band_role, session):
+def get_band_name(id, gender, band_role, session):
 	
-	if (session == "June"):
-		band_assignments = band_assignments_June
-	elif (session == "July"):
-		band_assignments = band_assignments_July
+	if (gender == "Male"):
+		if (session == "June"):
+			band_assignments = sec_Male_band_assignments_June
+		elif (session == "July"):
+			band_assignments = sec_Male_band_assignments_July
+		else:
+			band_assignments = sec_Male_band_assignments_August
 	else:
-		band_assignments = band_assignments_August
+		if (session == "June"):
+			band_assignments = sec_Female_band_assignments_June
+		elif (session == "July"):
+			band_assignments = sec_Female_band_assignments_July
+		else:
+			band_assignments = sec_Female_band_assignments_August
 
 	for key in band_assignments:
 		if (len(band_assignments[key][band_role]) != 0 and band_assignments[key][band_role][0][0] == id):
@@ -355,19 +373,26 @@ Manually inserts a given Camper into a given Band
 """
 def insert_into_band(id, band_role, gender, rank, session, band_name):
 
-	band_filename = "band_assignments_"+session+".p"
+	band_filename = "sec_"+gender+"_band_assignments_"+session+".p"
 
-	if (session == "June"):
-		band_assignments = band_assignments_June
-	elif (session == "July"):
-		band_assignments = band_assignments_July
+	if (gender == "Male"):
+		if (session == "June"):
+			band_assignments = sec_Male_band_assignments_June
+		elif (session == "July"):
+			band_assignments = sec_Male_band_assignments_July
+		else:
+			band_assignments = sec_Male_band_assignments_August
 	else:
-		band_assignments = band_assignments_August
+		if (session == "June"):
+			band_assignments = sec_Female_band_assignments_June
+		elif (session == "July"):
+			band_assignments = sec_Female_band_assignments_July
+		else:
+			band_assignments = sec_Female_band_assignments_August
 
 	# assuming the function band_insertion_check() was called before this function
 	band_assignments[band_name][band_role].append((id, gender, rank))
 	pickle.dump( band_assignments, open(band_filename, "wb") )
-
 
 
 """
@@ -375,12 +400,20 @@ Checks if a Camper could be inserted into a given 'band_name'
 """
 def band_insertion_check(id, band_role, gender, rank, session, band_name):
 
-	if (session == "June"):
-		band_assignments = band_assignments_June
-	elif (session == "July"):
-		band_assignments = band_assignments_July
+	if (gender == "Male"):
+		if (session == "June"):
+			band_assignments = sec_Male_band_assignments_June
+		elif (session == "July"):
+			band_assignments = sec_Male_band_assignments_July
+		else:
+			band_assignments = sec_Male_band_assignments_August
 	else:
-		band_assignments = band_assignments_August
+		if (session == "June"):
+			band_assignments = sec_Female_band_assignments_June
+		elif (session == "July"):
+			band_assignments = sec_Female_band_assignments_July
+		else:
+			band_assignments = sec_Female_band_assignments_August
 
 
 	if (len(band_assignments[band_name][band_role]) != 0):
@@ -393,19 +426,16 @@ def band_insertion_check(id, band_role, gender, rank, session, band_name):
 				curr_rank = temp_dict[i][0][2]
 				if curr_rank != "None": rank_count += 1
 
-		if(get_gender_distribution(gender, temp_dict)):
-			if (rank != "None" and rank_count < 4):
-				return 1
-			elif (rank == "None"):
-				return 1
-			else:
-				return 0
+		
+		if (rank != "None" and rank_count < 4):
+			return 1
+		elif (rank == "None"):
+			return 1
 		else:
-			return 0  
+			return 0
 
 
-
-def band_assignment_driver(camper_id_text, preferences):
+def sec_band_assignment_driver(camper_id_text, preferences):
 	c, conn = create_db()	
 	# get params from camper
 	c.execute("SELECT gender, band_role, camp_dates FROM application WHERE camper_id = '"+ camper_id_text + "'")
@@ -431,14 +461,14 @@ def band_assignment_driver(camper_id_text, preferences):
 
 	if(r1 and r1[0] == 0 and r2 and r2[0] == 0):
 		
-		band = assign_band(camper_id_text, band_role, gender, rank, session)
+		band = assign_sec_band(camper_id_text, band_role, gender, rank, session)
 		#insert into band DB
 		t = (camper_id_text, session, band, gender, int(rank))
-		c.execute("INSERT INTO band VALUES (?,?,?,?,?)", t)
+		c.execute("INSERT INTO second_band VALUES (?,?,?,?,?)", t)
 
 	# if the pref band mate has checked in, but not avoid band mate 
 	elif(r1 and r1[0] == 1 and r2 and r2[0] == 0):
-		c.execute("SELECT band FROM band WHERE camper_id = '"+ preferences[0].text() + "'")
+		c.execute("SELECT band FROM second_band WHERE camper_id = '"+ preferences[0].text() + "'")
 		r4 = c.fetchone()
 		pref_band = r4[0]
 
@@ -446,29 +476,29 @@ def band_assignment_driver(camper_id_text, preferences):
 			insert_into_band(camper_id_text, band_role, gender, rank, session, pref_band)
 			# insert into band DB
 			t = (camper_id_text, session, pref_band, gender, int(rank))
-			c.execute("INSERT INTO band VALUES (?,?,?,?,?)", t)
+			c.execute("INSERT INTO second_band VALUES (?,?,?,?,?)", t)
 		else:
-			band = assign_band(camper_id_text, band_role, gender, rank, session)
+			band = assign_sec_band(camper_id_text, band_role, gender, rank, session)
 			# insert into band DB
 			t = (camper_id_text, session, band, gender, int(rank))
-			c.execute("INSERT INTO band VALUES (?,?,?,?,?)", t)
+			c.execute("INSERT INTO second_band VALUES (?,?,?,?,?)", t)
 
 	elif(r2 and r2[0] == 1 and r1 and r1[0] == 0):
-		c.execute("SELECT band FROM band WHERE camper_id = '"+ preferences[1].text() + "'")
+		c.execute("SELECT band FROM second_band WHERE camper_id = '"+ preferences[1].text() + "'")
 		r4 = c.fetchone()
 		band_avoid = r4[0]
 
-		band = assign_band(camper_id_text, band_role, gender, rank, session, avoid_band=band_avoid)
+		band = assign_sec_band(camper_id_text, band_role, gender, rank, session, avoid_band=band_avoid)
 		# insert into band DB
 		t = (camper_id_text, session, band, gender, int(rank))
-		c.execute("INSERT INTO band VALUES (?,?,?,?,?)", t)
+		c.execute("INSERT INTO second_band VALUES (?,?,?,?,?)", t)
 
 	elif(r2 and r2[0] == 1 and r1 and r1[0] == 1):
-		c.execute("SELECT band FROM band WHERE camper_id = '"+ preferences[0].text() + "'")
+		c.execute("SELECT band FROM second_band WHERE camper_id = '"+ preferences[0].text() + "'")
 		r4 = c.fetchone()
 		pref_band = r4[0]
 
-		c.execute("SELECT band FROM band WHERE camper_id = '"+ preferences[1].text() + "'")
+		c.execute("SELECT band FROM second_band WHERE camper_id = '"+ preferences[1].text() + "'")
 		r5 = c.fetchone()
 		band_avoid = r5[0]
 
@@ -476,47 +506,25 @@ def band_assignment_driver(camper_id_text, preferences):
 			insert_into_band(camper_id_text, band_role, gender, rank, session, pref_band)
 			# insert into band DB
 			t = (camper_id_text, session, pref_band, gender, int(rank))
-			c.execute("INSERT INTO band VALUES (?,?,?,?,?)", t)
+			c.execute("INSERT INTO second_band VALUES (?,?,?,?,?)", t)
 		else:
-			band = assign_band(camper_id_text, band_role, gender, rank, session, avoid_band=band_avoid)
+			band = assign_sec_band(camper_id_text, band_role, gender, rank, session, avoid_band=band_avoid)
 			# insert into band DB
 			t = (camper_id_text, session, band, gender, int(rank))
-			c.execute("INSERT INTO band VALUES (?,?,?,?,?)", t)
+			c.execute("INSERT INTO second_band VALUES (?,?,?,?,?)", t)
 
 
 	else:
-		band = assign_band(camper_id_text, band_role, gender, rank, session)
+		band = assign_sec_band(camper_id_text, band_role, gender, rank, session)
 		#insert into band DB
 		t = (camper_id_text, session, band, gender, int(rank))
-		c.execute("INSERT INTO band VALUES (?,?,?,?,?)", t)
+		c.execute("INSERT INTO second_band VALUES (?,?,?,?,?)", t)
 
 	save_db_changes(conn)
 
 
 
-def swap_request(id1, band_role1, id2, band_role2):
-
-	if band_role1 == band_role2: return 0 
-
-
-	# find the bands for id1 and id2 
-	for key in band_assignments:
-
-		if( band_assignments[key][band_role1][0][0] == id1):
-			id1_band = key
-
-		if( band_assignments[key][band_role2][0][0] == id2):
-			id2_band = key
-
-	if id1_band == id2_band: return 0 
 
 
 
-
-"""
-assign_band("105", "Drummer", "Male", 3, "July", avoid_band="Band1")
-print(band_assignments_July)
-print(band_assignments_June)
-
-#"""
