+++ sd_dorm_assign.py
@@ -1,9 +1,8 @@
 
-
+from sqlite_dbv3 import * 
 import pickle
 import os 
 from collections import Counter
-
 
 
 
@@ -12,9 +11,9 @@ Check if pickle files containing the saved Dorm dictionaries exist. If so, load 
 
 """
 
-if (os.path.exists("male_dorms.p")):
+if (os.path.exists("June_male_dorms.p")):
 	
-	male_dorms = pickle.load(open("male_dorms.p", "rb"))
+	June_male_dorms = pickle.load(open("June_male_dorms.p", "rb"))
 
 else:
 	
@@ -25,7 +24,7 @@ else:
 
 	"""
 
-	male_dorms = {
+	June_male_dorms = {
 	
 		'A' : [],
 		'B': [],
@@ -35,13 +34,13 @@ else:
 
 
 
-if (os.path.exists("female_dorms.p")):
+if (os.path.exists("June_female_dorms.p")):
 	
-	female_dorms = pickle.load(open("female_dorms.p", "rb"))
+	June_female_dorms = pickle.load(open("June_female_dorms.p", "rb"))
 
 else:
 	
-	female_dorms = {
+	June_female_dorms = {
 	
 		'D' : [],
 		'E': [],
@@ -49,7 +48,79 @@ else:
 
 	} 
 
+if (os.path.exists("July_male_dorms.p")):
+	
+	July_male_dorms = pickle.load(open("July_male_dorms.p", "rb"))
 
+else:
+	
+	"""
+	
+	Each Dorm is saved in dictionary according to gender
+	For example, the male dorm 'A' is daved as: 'A' : [ (Camper_ID, Age), (Camper_ID, Age), .. ] 
+
+	"""
+
+	July_male_dorms = {
+	
+		'A' : [],
+		'B': [],
+		'C' : []
+
+	} 
+
+
+
+if (os.path.exists("July_female_dorms.p")):
+	
+	July_female_dorms = pickle.load(open("July_female_dorms.p", "rb"))
+
+else:
+	
+	July_female_dorms = {
+	
+		'D' : [],
+		'E': [],
+		'F' : []
+
+	} 
+
+if (os.path.exists("August_male_dorms.p")):
+	
+	August_male_dorms = pickle.load(open("August_male_dorms.p", "rb"))
+
+else:
+	
+	"""
+	
+	Each Dorm is saved in dictionary according to gender
+	For example, the male dorm 'A' is daved as: 'A' : [ (Camper_ID, Age), (Camper_ID, Age), .. ] 
+
+	"""
+
+	August_male_dorms = {
+	
+		'A' : [],
+		'B': [],
+		'C' : []
+
+	} 
+
+
+
+if (os.path.exists("August_female_dorms.p")):
+	
+	August_female_dorms = pickle.load(open("August_female_dorms.p", "rb"))
+
+else:
+	
+	August_female_dorms = {
+	
+		'D' : [],
+		'E': [],
+		'F' : []
+
+	} 
 
 """
 Given a age (13-18), the approrpiate 'range' is calculated
@@ -72,6 +143,80 @@ def get_age_range(age):
 
 	return a,b 
 
+"""
+Obtain the Dorm Number of a given ID
+
+"""
+def get_dorm_num(id, gender, session):
+
+	if gender == 'Male':
+
+		if (session == "June"):
+			male_dorms = June_male_dorms
+		elif (session == "July"):
+			male_dorms = July_male_dorms
+		else:
+			male_dorms = August_male_dorms
+
+
+		for i in male_dorms:
+			for (j,k) in male_dorms[i]:
+				if str(j) == str(id):
+					return i
+		return "None"
+
+	else:
+
+		if (session == "June"):
+			female_dorms = June_female_dorms
+		elif (session == "July"):
+			female_dorms = July_female_dorms
+		else:
+			female_dorms = August_female_dorms
+
+
+		for i in female_dorms:
+			for (j,k) in female_dorms[i]:
+				if str(j) == str(id):
+					return i
+		return "None"
+
+"""
+Manually insert Camper into a Dorm (after checking if ok, hopefully)
+
+"""
+def insert_into_dorm(id, gender, age, session, dorm_name):
+
+	
+	if gender == "Male":
+
+		filename = session + "_male_dorms.p"
+
+		if (session == "June"):
+			male_dorms = June_male_dorms
+		elif (session == "July"):
+			male_dorms = July_male_dorms
+		else:
+			male_dorms = August_male_dorms
+
+		male_dorms[dorm_name].append((id,age))
+		pickle.dump( male_dorms, open( filename, "wb" ) )
+
+	else:
+
+		filename = session + "_female_dorms.p"
+
+		if (session == "June"):
+			female_dorms = June_female_dorms
+		elif (session == "July"):
+			female_dorms = July_female_dorms
+		else:
+			female_dorms = August_female_dorms
+
+		female_dorms[dorm_name].append((id,age))
+		pickle.dump( female_dorms, open( filename, "wb" ) )
+
+
 
 """
 
@@ -82,7 +227,17 @@ age: Camper age
 a,b : Camper age range (e.g. if the camper is 16 yrs old, then the age range will be a=15, b=16)
 
 """
-def assign_male(id, age, a, b):
+def assign_male(id, age, a, b, session, dorm_avoid=None):
+
+	filename = session + "_male_dorms.p"
+	
+	if (session == "June"):
+		male_dorms = June_male_dorms
+	elif (session == "July"):
+		male_dorms = July_male_dorms
+	else:
+		male_dorms = August_male_dorms
+
 
 	for key in male_dorms:
 		arr = male_dorms[key]
@@ -102,13 +257,13 @@ def assign_male(id, age, a, b):
 		print("TEST: length", len(arr))
 		"""
 
-		if( len(arr) < 8 and sum_age_count < 3):
+		if( len(arr) < 8 and sum_age_count < 3 and key != dorm_avoid ):
 			# then, we could add this camper to our dorm
 			male_dorms[key].append((id,age))
-			pickle.dump( male_dorms, open( "male_dorms.p", "wb" ) )
+			pickle.dump( male_dorms, open( filename, "wb" ) )
 
 			# TODO: add in SQL INSERT command to insert camper into DORM TABLE
-			return 1
+			return key 
 
 
 
@@ -121,7 +276,16 @@ age : Camper age
 a,b : Camper age range (e.g. if the camper is 16 yrs old, then the age range will be a=15, b=16)
 
 """
-def assign_female(id, age, a, b):
+def assign_female(id, age, a, b, session, dorm_avoid=None):
+
+	filename = session + "_female_dorms.p"
+	
+	if (session == "June"):
+		female_dorms = June_female_dorms
+	elif (session == "July"):
+		female_dorms = July_female_dorms
+	else:
+		female_dorms = August_female_dorms
 
 	for key in female_dorms:
 		arr = female_dorms[key]
@@ -130,13 +294,13 @@ def assign_female(id, age, a, b):
 		if (age_count[str(a)]): sum_age_count += age_count[str(a)]
 		if (age_count[str(b)]): sum_age_count += age_count[str(b)]
 
-		if( len(arr) < 8 and sum_age_count < 3):
+		if( len(arr) < 8 and sum_age_count < 3 and key != dorm_avoid):
 			# then, we could add this camper to our dorm
 			female_dorms[key].append((id,int(age)))
-			pickle.dump( female_dorms, open( "female_dorms.p", "wb" ) )
+			pickle.dump( female_dorms, open( filename, "wb" ) )
 
 			# TODO: add in SQL INSERT command to insert camper into DORM TABLE
-			return 1
+			return key 
 
 
 
@@ -148,17 +312,24 @@ gender : Camper's Gender
 age : Camper age
 
 """
-def assign_dorm(id, gender, age):
+def assign_dorm(id, gender, age, session, dorm_avoid=None):
 
 	a,b = get_age_range(int(age))
 	#print("TEST: range: ", a, b)
 
-	if (gender == "male" or gender == "m" or gender == "Male" or gender == "M"):
-		assign_male(id, age, a, b)
+	if (gender == "Male"):
+		if dorm_avoid:
+			dorm = assign_male(id, age, a, b, session, dorm_avoid=dorm_avoid)
+		else:
+			dorm = assign_male(id, age, a, b, session)
 
 	else:
-		assign_female(id, age, a, b)
+		if dorm_avoid:
+			dorm = assign_female(id, age, a, b, session, dorm_avoid=dorm_avoid)
+		else:
+			dorm = assign_female(id, age, a, b, session)
 
+	return dorm 
 
 
 """
@@ -175,33 +346,53 @@ Returns 1 if it is ok to insert that Camper into the indicated Dorm
 Returns 0 if not
 
 """
-def dorm_insertion_check(id, gender, age, dorm_name):
+def dorm_insertion_check(id, gender, age, session, dorm_name):
 	
 	a,b = get_age_range(int(age))
 
-	if gender == "male" and dorm_name in list(male_dorms.keys()):
-		
-		arr = male_dorms[dorm_name]
-		age_count = Counter(elt[1] for elt in arr)
-		sum_age_count = 0
-		if (age_count[str(a)]): sum_age_count += age_count[str(a)]
-		if (age_count[str(b)]): sum_age_count += age_count[str(b)]
+	if gender == "Male":
 
-		if ( len(arr) < 8 and sum_age_count < 3 ):
-			return 1
+		if (session == "June"):
+			male_dorms = June_male_dorms
+		elif (session == "July"):
+			male_dorms = July_male_dorms
+		else:
+			male_dorms = August_male_dorms	
+				
+
+		if dorm_name in list(male_dorms.keys()):
+			
+			arr = male_dorms[dorm_name]
+			age_count = Counter(elt[1] for elt in arr)
+			sum_age_count = 0
+			if (age_count[str(a)]): sum_age_count += age_count[str(a)]
+			if (age_count[str(b)]): sum_age_count += age_count[str(b)]
+
+			if ( len(arr) < 8 and sum_age_count < 3 ):
+				return 1
 
 		return 0 
 
-	elif gender == "female" and dorm_name in list(female_dorms.keys()):
+	elif gender == "Female":
 
-		arr = female_dorms[dorm_name]
-		age_count = Counter(elt[1] for elt in arr)
-		sum_age_count = 0
-		if (age_count[str(a)]): sum_age_count += age_count[str(a)]
-		if (age_count[str(b)]): sum_age_count += age_count[str(b)]
+		if (session == "June"):
+			female_dorms = June_female_dorms
+		elif (session == "July"):
+			female_dorms = July_female_dorms
+		else:
+			female_dorms = August_female_dorms	
 
-		if ( len(arr) < 8 and sum_age_count < 3 ):
-			return 1
+
+		if dorm_name in list(female_dorms.keys()):
+
+			arr = female_dorms[dorm_name]
+			age_count = Counter(elt[1] for elt in arr)
+			sum_age_count = 0
+			if (age_count[str(a)]): sum_age_count += age_count[str(a)]
+			if (age_count[str(b)]): sum_age_count += age_count[str(b)]
+
+			if ( len(arr) < 8 and sum_age_count < 3 ):
+				return 1
 
 		return 0
 
@@ -210,12 +401,97 @@ def dorm_insertion_check(id, gender, age, dorm_name):
 		return 0 
 
 
+def dorm_assignment_driver(camper_id_text, preferences):
+	# Assign Camper a Dorm 
+	
+	c, conn = create_db()	
+	# get params from camper
+	c.execute("SELECT gender, age, camp_dates FROM application WHERE camper_id = '"+ camper_id_text + "'")
+	r0 = c.fetchone() 
+
+	# Check if roomate_pref has checked in
+	c.execute("SELECT has_checked_in FROM check_in WHERE camper_id = '"+ preferences[0].text() + "'")
+	r1 = c.fetchone()
+
+	# Check if roomate_avoid has checked in 
+	c.execute("SELECT has_checked_in FROM check_in WHERE camper_id = '"+ preferences[1].text() + "'")
+	r2 = c.fetchone()
+
+	if(r1 and r1[0] == 0 and r2 and r2[0] == 0):
+		dorm = assign_dorm(camper_id_text, r0[0], r0[1], r0[2])
+		
+		# INSERT INTO dorm table
+		t = (camper_id_text, r0[2], dorm, r0[0], int(r0[1]))
+		c.execute("INSERT INTO dorm VALUES (?,?,?,?,?)", t)
+ 
+
+	# if the pref roomate has checked in, but not roomate_avoid
+	elif(r1 and r1[0] == 1 and r2 and r2[0] == 0):
+
+		c.execute("SELECT dorm_number FROM dorm WHERE camper_id = '"+ preferences[0].text() + "'")
+		r4 = c.fetchone()
+		pref_dorm = r4[0]
+
+		if(dorm_insertion_check(camper_id_text, r0[0], r0[1], r0[2], pref_dorm)):
+			insert_into_dorm(camper_id_text, r0[0], r0[1], r0[2], pref_dorm)
+			
+			# INSERT INTO dorm table
+			t = (camper_id_text, r0[2], pref_dorm, r0[0], int(r0[1]))
+			c.execute("INSERT INTO dorm VALUES (?,?,?,?,?)", t)
+
+		else:
+			dorm = assign_dorm(camper_id_text, r0[0], r0[1], r0[2])
+			# INSERT INTO dorm table 
+			t = (camper_id_text, r0[2], dorm, r0[0], int(r0[1]))
+			c.execute("INSERT INTO dorm VALUES (?,?,?,?,?)", t)
+
+	# if the roomate_avoid has checked in, but not roomate_pref	
+	elif(r2 and r2[0] == 1 and r1 and r1[0] == 0 ):
+		c.execute("SELECT dorm_number FROM dorm WHERE camper_id = '"+ preferences[1].text() + "'")
+		r4 = c.fetchone()
+		dorm_avoid = r4[0]
+
+		dorm = assign_dorm(camper_id_text, r0[0], r0[1], r0[2], dorm_avoid=dorm_avoid)
+		# INSERT INTO dorm table
+		t = (camper_id_text, r0[2], dorm, r0[0], int(r0[1]))
+		c.execute("INSERT INTO dorm VALUES (?,?,?,?,?)", t) 
+
+	elif(r2 and r2[0] == 1 and r1 and r1[0] == 1):
+		c.execute("SELECT dorm_number FROM dorm WHERE camper_id = '"+ preferences[0].text() + "'")
+		r4 = c.fetchone()
+		pref_dorm = r4[0]
+
+		c.execute("SELECT dorm_number FROM dorm WHERE camper_id = '"+ preferences[1].text() + "'")
+		r5 = c.fetchone()
+		dorm_avoid = r5[0]
+
+		if(dorm_insertion_check(camper_id_text, r0[0], r0[1], r0[2], pref_dorm)):
+			insert_into_dorm(camper_id_text, r0[0], r0[1], r0[2], pref_dorm)
+			# INSERT INTO dorm table
+			t = (camper_id_text, r0[2], pref_dorm, r0[0], int(r0[1]))
+			c.execute("INSERT INTO dorm VALUES (?,?,?,?,?)", t)
+		else:
+			dorm = assign_dorm(camper_id_text, r0[0], r0[1], r0[2], dorm_avoid=dorm_avoid)
+			# INSERT INTO dorm table
+			t = (camper_id_text, r0[2], dorm, r0[0], int(r0[1]))
+			c.execute("INSERT INTO dorm VALUES (?,?,?,?,?)", t) 
+
+	else:
+		#do regular assignement
+		dorm = assign_dorm(camper_id_text, r0[0], r0[1], r0[2])
+		# INSERT INTO dorm table
+		t = (camper_id_text, r0[2], dorm, r0[0], int(r0[1]))
+		c.execute("INSERT INTO dorm VALUES (?,?,?,?,?)", t)
+
+	save_db_changes(conn)  
+
+
 
 
 """
-assign_dorm("101","male","18")
-print(male_dorms)
-print(female_dorms)
+assign_dorm("101","Male","18","July")
+print(July_male_dorms)
+print(July_female_dorms)
 """
